Reformat and move more files
