+++ DataConverter.Test/StructureTests/GameVersionConverterTest.cs
@@ -3,6 +3,7 @@ using DataConverter.Converters;
 using FluentAssertions;
 using NUnit.Framework;
 using WoWsShipBuilder.DataStructures;
+using WoWsShipBuilder.DataStructures.Versioning;
 
 namespace DataConverter.Test.StructureTests;
