+++ DataConverter.Test/StructureTests/GameVersionTest.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
 using WoWsShipBuilder.DataStructures;
+using WoWsShipBuilder.DataStructures.Versioning;
 
 namespace DataConverter.Test.StructureTests
 {
