+++ DataConverter/Services/IVersionCheckService.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using DataConverter.Data;
 using WoWsShipBuilder.DataStructures;
+using WoWsShipBuilder.DataStructures.Versioning;
 
 namespace DataConverter.Services;
