+++ WoWsShipBuilder.DataStructures/Module.cs
@@ -1,12 +1,13 @@
-﻿namespace WoWsShipBuilder.DataStructures
+﻿namespace WoWsShipBuilder.DataStructures;
+
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable CollectionNeverUpdated.Global
+public class Module
 {
-    public class Module
-    {
-        public int CostCr { get; set; }
-        public int CostXp { get; set; }
-        public long Id { get; set; }
-        public string Index { get; set; }
-        public string Name { get; set; }
-        public ModuleType Type { get; set; }
-    }
+    public int CostCr { get; set; }
+    public int CostXp { get; set; }
+    public long Id { get; set; }
+    public string Index { get; set; } = string.Empty;
+    public string Name { get; set; } = string.Empty;
+    public ModuleType Type { get; set; }
 }
