--- WoWsShipBuilder.DataStructures/ReplayVersionDetails.cs
@@ -1,4 +1,0 @@
-﻿namespace WoWsShipBuilder.DataStructures
-{
-    public record ReplayVersionDetails(byte OnChatMessage, byte OnArenaStateReceived);
-}
