+++ WoWsShipBuilder.DataStructures/ShipSummary.cs
@@ -1,8 +1,7 @@
-﻿#nullable enable
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
-namespace WoWsShipBuilder.DataStructures
-{
-    public sealed record ShipSummary(string Index, Nation Nation, int Tier, ShipClass ShipClass, ShipCategory Category, string? PrevShipIndex, List<string>? NextShipsIndex = null);
-}
-#nullable restore
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable CollectionNeverUpdated.Global
+namespace WoWsShipBuilder.DataStructures;
+
+public sealed record ShipSummary(string Index, Nation Nation, int Tier, ShipClass ShipClass, ShipCategory Category, string? PrevShipIndex, List<string>? NextShipsIndex = null);
