+++ WoWsShipBuilder.DataStructures/ShipTurretOverride.cs
@@ -1,12 +1,13 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
-#nullable enable
-namespace WoWsShipBuilder.DataStructures
-{
-    /// <summary>
-    /// Container class for turret position overrides.
-    /// </summary>
-    /// <param name="ArtilleryTurretOverrides">Maps the name of an artillery module to a dictionary mapping the turret keys to their position overrides.</param>
-    public sealed record ShipTurretOverride(Dictionary<string, Dictionary<string, TurretOrientation>> ArtilleryTurretOverrides);
-}
-#nullable restore
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable CollectionNeverUpdated.Global
+namespace WoWsShipBuilder.DataStructures;
+
+/// <summary>
+/// Container class for turret position overrides.
+/// </summary>
+/// <param name="ArtilleryTurretOverrides">Maps the name of an artillery module to a dictionary mapping the turret keys to their position overrides.</param>
+[Obsolete("No longer used", true)]
+public sealed record ShipTurretOverride(Dictionary<string, Dictionary<string, TurretOrientation>> ArtilleryTurretOverrides);
